refactor: Add a return value to the _create_bundle callback
In anticipation of
https://github.com/aws-deadline/deadline-cloud/pull/673,
"fix!: Implement improved job attachments containment warning."
The return value provides the parameter values for use as known paths in
the warning message.

This change is backwards-compatible, because before that PR, the return
value is ignored, and after that PR, the return value will be used to
improve the message.

Signed-off-by: Mark Wiebe <399551+mwiebe@users.noreply.github.com>